Ajustes nas procedures e lógica de solicitar inspeção

+++ ApiIntegracaoConfitec/Interfaces/Infrastructure/Repository/ISompoRepository.cs
@@ -1,6 +1,7 @@
 ﻿using ApiIntegracaoConfitec.Models.Confitec;
 using ApiIntegracaoConfitec.Models.Domain.Handler;
 using ApiIntegracaoConfitec.Models.Entity;
+using System;
 using System.Threading.Tasks;
 
 namespace ApiIntegracaoConfitec.Interfaces.Infrastructure.Repository
@@ -8,7 +9,7 @@ namespace ApiIntegracaoConfitec.Interfaces.Infrastructure.Repository
     public interface ISompoRepository
     {
         // Solicitar inspeção
-        Task<DadosInspecao> RetornarDadosInspecao(int Num_PI, int Num_Local, int Tip_Emissao);
+        Task<DadosInspecao> RetornarDadosInspecao(Int64 Num_PI, int Num_Local, int Tip_Emissao);
         Task<DadosAutenticacao> RetornarDadosAutenticacao();
         Task<QueryResult> GravarRetornoSolicitarInspecao(ConfitecSolicitarInspecao responseSolicitarInspecao);
         // Gravar laudo
