+++ ApiIntegracaoConfitec/Models/Domain/Handler/BuscarDadosSolicitarInspecaoRequest.cs
@@ -1,5 +1,6 @@
 ﻿using ApiIntegracaoConfitec.Domain.Utility;
 using ApiIntegracaoConfitec.Helpers;
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace ApiIntegracaoConfitec.Models.Domain.Handler
@@ -7,8 +8,8 @@ namespace ApiIntegracaoConfitec.Models.Domain.Handler
     public class BuscarDadosSolicitarInspecaoRequest
     {
         [Required(ErrorMessage = "Número do PI não informado")]
-        [Range(1, int.MaxValue, ErrorMessage = "Informe um Número PI maior que zero")]
-        public int Num_PI { get; set; }
+        [Range(1, Int64.MaxValue, ErrorMessage = "Informe um Número PI maior que zero")]
+        public Int64 Num_PI { get; set; }
 
         [Required(ErrorMessage = "Número do Local não informado")]
         [Range(1, int.MaxValue, ErrorMessage = "Informe um Número Local maior que zero")]
@@ -18,7 +19,7 @@ namespace ApiIntegracaoConfitec.Models.Domain.Handler
         [Range(1, int.MaxValue, ErrorMessage = "Informe um Tipo de Emissao maior que zero")]
         public int Tip_Emissao { get; set; }
 
-        public BuscarDadosSolicitarInspecaoRequest(int Num_PI, int Num_Local, int Tip_Emissao)
+        public BuscarDadosSolicitarInspecaoRequest(Int64 Num_PI, int Num_Local, int Tip_Emissao)
         {
             this.Num_PI = Num_PI;
             this.Num_Local = Num_Local;
